Split view numba functions into separate file
