+++ pysheds/sview.py
@@ -1,11 +1,9 @@
 import numpy as np
-from scipy import spatial
-from scipy import interpolate
-from numba import njit, prange
-from numba.types import float64, UniTuple
 import pyproj
 from affine import Affine
 from distutils.version import LooseVersion
+
+import pysheds._sview as _self
 
 _OLD_PYPROJ = LooseVersion(pyproj.__version__) < LooseVersion('2.2')
 _pyproj_init = '+init=epsg:4326' if _OLD_PYPROJ else 'epsg:4326'
@@ -62,11 +60,6 @@ class Raster(np.ndarray):
         bbox = self.viewfinder.bbox
         extent = (bbox[0], bbox[2], bbox[1], bbox[3])
         return extent
-    @property
-    def cellsize(self):
-        dy, dx = self.dy_dx
-        cellsize = (dy + dx) / 2
-        return cellsize
     @property
     def affine(self):
         return self.viewfinder.affine
@@ -215,7 +208,7 @@ class ViewFinder():
         target_view = self
         return View.view(raster, data_view, target_view, interpolation='nearest')
 
-    def grid_indices(self, affine=None, shape=None, col_ascending=True, row_ascending=False):
+    def grid_indices(self, affine=None, shape=None):
         """
         Return row and column coordinates of a bounding box at a
         given cellsize.
@@ -234,33 +227,9 @@ class ViewFinder():
             shape = self.shape
         y_ix = np.arange(shape[0])
         x_ix = np.arange(shape[1])
-        if row_ascending:
-            y_ix = y_ix[::-1]
-        if not col_ascending:
-            x_ix = x_ix[::-1]
         x, _ = affine * np.vstack([x_ix, np.zeros(shape[1])])
         _, y = affine * np.vstack([np.zeros(shape[0]), y_ix])
         return y, x
-
-    def move_window(self, dxmin, dymin, dxmax, dymax):
-        """
-        Move bounding box window by integer indices
-        """
-        cell_height, cell_width  = self.dy_dx
-        nrows_old, ncols_old = self.shape
-        xmin_old, ymin_old, xmax_old, ymax_old = self.bbox
-        new_bbox = (xmin_old + dxmin*cell_width, ymin_old + dymin*cell_height,
-                    xmax_old + dxmax*cell_width, ymax_old + dymax*cell_height)
-        new_shape = (nrows_old + dymax - dymin,
-                     ncols_old + dxmax - dxmin)
-        new_mask = np.ones(new_shape).astype(bool)
-        mask_values = self._mask[max(dymin, 0):min(nrows_old + dymax, nrows_old),
-                                 max(dxmin, 0):min(ncols_old + dxmax, ncols_old)]
-        new_mask[max(0, dymax):max(0, dymax) + mask_values.shape[0],
-                 max(0, -dxmin):max(0, -dxmin) + mask_values.shape[1]] = mask_values
-        self.bbox = new_bbox
-        self.shape = new_shape
-        self.mask = new_mask
 
 class View():
     def __init__(self):
@@ -307,6 +276,29 @@ class View():
             out.metadata.update(data.metadata)
         out.metadata.update(new_metadata)
         return out
+
+    @classmethod
+    def affine_transform(cls, affine, x, y):
+        # Check affine input type
+        try:
+            assert isinstance(affine, Affine)
+            affine = tuple(affine)
+        except:
+            raise TypeError('`affine` must be an Affine instance')
+        # Vector case
+        if hasattr(x, '__len__'):
+            if hasattr(y, '__len__'):
+                x = np.asarray(x).astype(np.float64)
+                y = np.asarray(y).astype(np.float64)
+                x_t, y_t = _self._affine_map_vec_numba(affine, x, y)
+            else:
+                raise TypeError('If `x` is a sequence, `y` must also be a sequence')
+        # Scalar case
+        else:
+            x = float(x)
+            y = float(y)
+            x_t, y_t = _self._affine_map_scalar_numba(affine, x, y)
+        return x_t, y_t
 
     @classmethod
     def trim_zeros(cls, data, pad=(0,0,0,0)):
@@ -435,17 +427,17 @@ class View():
     @classmethod
     def _view_same_crs(cls, view, data, data_view, target_view, interpolation='nearest'):
         y, x = target_view.axes
-        inv_affine = tuple(~data_view.affine)
-        _, y_ix = affine_map(inv_affine,
-                            np.zeros(target_view.shape[0], dtype=np.float64),
-                            y)
-        x_ix, _ = affine_map(inv_affine,
-                            x,
-                            np.zeros(target_view.shape[1], dtype=np.float64))
+        inv_affine = ~data_view.affine
+        _, y_ix = cls.affine_transform(inv_affine,
+                                       np.zeros(target_view.shape[0],
+                                                dtype=np.float64), y)
+        x_ix, _ = cls.affine_transform(inv_affine, x,
+                                       np.zeros(target_view.shape[1],
+                                                dtype=np.float64))
         if interpolation == 'nearest':
-            view = _view_fill_by_axes_nearest_numba(data, view, y_ix, x_ix)
+            view = _self._view_fill_by_axes_nearest_numba(data, view, y_ix, x_ix)
         elif interpolation == 'linear':
-            view = _view_fill_by_axes_linear_numba(data, view, y_ix, x_ix)
+            view = _self._view_fill_by_axes_linear_numba(data, view, y_ix, x_ix)
         else:
             raise ValueError('Interpolation method must be one of: `nearest`, `linear`')
         return view
@@ -455,141 +447,13 @@ class View():
         y, x = target_view.coords.T
         xt, yt = pyproj.transform(target_view.crs, data_view.crs, x=x, y=y,
                                   errcheck=True, always_xy=True)
-        inv_affine = tuple(~data_view.affine)
-        x_ix, y_ix = affine_map(inv_affine, xt, yt)
+        inv_affine = ~data_view.affine
+        x_ix, y_ix = cls.affine_transform(inv_affine, xt, yt)
         if interpolation == 'nearest':
-            view = _view_fill_by_entries_nearest_numba(data, view, y_ix, x_ix)
+            view = _self._view_fill_by_entries_nearest_numba(data, view, y_ix, x_ix)
         elif interpolation == 'linear':
-            view = _view_fill_by_entries_linear_numba(data, view, y_ix, x_ix)
+            view = _self._view_fill_by_entries_linear_numba(data, view, y_ix, x_ix)
         else:
             raise ValueError('Interpolation method must be one of: `nearest`, `linear`')
         return view
 
-@njit(parallel=True)
-def _view_fill_numba(data, out, y_ix, x_ix, y_passed, x_passed):
-    n = x_ix.size
-    m = y_ix.size
-    for i in prange(m):
-        for j in prange(n):
-            if (y_passed[i]) & (x_passed[j]):
-                out[i, j] = data[y_ix[i], x_ix[j]]
-    return out
-
-@njit(parallel=True)
-def _view_fill_by_axes_nearest_numba(data, out, y_ix, x_ix):
-    m, n = y_ix.size, x_ix.size
-    M, N = data.shape
-    # Currently need to use inplace form of round
-    y_near = np.empty(m, dtype=np.int64)
-    x_near = np.empty(n, dtype=np.int64)
-    np.around(y_ix, 0, y_near).astype(np.int64)
-    np.around(x_ix, 0, x_near).astype(np.int64)
-    y_in_bounds = ((y_near >= 0) & (y_near < M))
-    x_in_bounds = ((x_near >= 0) & (x_near < N))
-    for i in prange(m):
-        for j in prange(n):
-            if (y_in_bounds[i]) & (x_in_bounds[j]):
-                out[i, j] = data[y_near[i], x_near[j]]
-    return out
-
-@njit(parallel=True)
-def _view_fill_by_axes_linear_numba(data, out, y_ix, x_ix):
-    m, n = y_ix.size, x_ix.size
-    M, N = data.shape
-    # Find which cells are in bounds
-    y_in_bounds = ((y_ix >= 0) & (y_ix < M))
-    x_in_bounds = ((x_ix >= 0) & (x_ix < N))
-    # Compute upper and lower values of y and x
-    y_floor = np.floor(y_ix).astype(np.int64)
-    y_ceil = y_floor + 1
-    x_floor = np.floor(x_ix).astype(np.int64)
-    x_ceil = x_floor + 1
-    # Compute fractional distance between adjacent cells
-    ty = (y_ix - y_floor)
-    tx = (x_ix - x_floor)
-    # Handle lower and right boundaries
-    lower_boundary = (y_ceil == M)
-    right_boundary = (x_ceil == N)
-    y_ceil[lower_boundary] = y_floor[lower_boundary]
-    x_ceil[right_boundary] = x_floor[right_boundary]
-    ty[lower_boundary] = 0.
-    tx[right_boundary] = 0.
-    for i in prange(m):
-        for j in prange(n):
-            if (y_in_bounds[i]) & (x_in_bounds[j]):
-                ul = data[y_floor[i], x_floor[j]]
-                ur = data[y_floor[i], x_ceil[j]]
-                ll = data[y_ceil[i], x_floor[j]]
-                lr = data[y_ceil[i], x_ceil[j]]
-                value = ( ( ( 1 - tx[j] ) * ( 1 - ty[i] ) * ul )
-                         + ( tx[j] * ( 1 - ty[i] ) * ur )
-                         + ( ( 1 - tx[j] ) * ty[i] * ll )
-                         + ( tx[j] * ty[i] * lr ) )
-                out[i, j] = value
-    return out
-
-@njit(parallel=True)
-def _view_fill_by_entries_nearest_numba(data, out, y_ix, x_ix):
-    m, n = y_ix.size, x_ix.size
-    M, N = data.shape
-    # Currently need to use inplace form of round
-    y_near = np.empty(m, dtype=np.int64)
-    x_near = np.empty(n, dtype=np.int64)
-    np.around(y_ix, 0, y_near).astype(np.int64)
-    np.around(x_ix, 0, x_near).astype(np.int64)
-    y_in_bounds = ((y_near >= 0) & (y_near < M))
-    x_in_bounds = ((x_near >= 0) & (x_near < N))
-    # x and y indices should be the same size
-    assert(n == m)
-    for i in prange(n):
-        if (y_in_bounds[i]) & (x_in_bounds[i]):
-            out.flat[i] = data[y_near[i], x_near[i]]
-    return out
-
-@njit(parallel=True)
-def _view_fill_by_entries_linear_numba(data, out, y_ix, x_ix):
-    m, n = y_ix.size, x_ix.size
-    M, N = data.shape
-    # Find which cells are in bounds
-    y_in_bounds = ((y_ix >= 0) & (y_ix < M))
-    x_in_bounds = ((x_ix >= 0) & (x_ix < N))
-    # Compute upper and lower values of y and x
-    y_floor = np.floor(y_ix).astype(np.int64)
-    y_ceil = y_floor + 1
-    x_floor = np.floor(x_ix).astype(np.int64)
-    x_ceil = x_floor + 1
-    # Compute fractional distance between adjacent cells
-    ty = (y_ix - y_floor)
-    tx = (x_ix - x_floor)
-    # Handle lower and right boundaries
-    lower_boundary = (y_ceil == M)
-    right_boundary = (x_ceil == N)
-    y_ceil[lower_boundary] = y_floor[lower_boundary]
-    x_ceil[right_boundary] = x_floor[right_boundary]
-    ty[lower_boundary] = 0.
-    tx[right_boundary] = 0.
-    # x and y indices should be the same size
-    assert(n == m)
-    for i in prange(n):
-        if (y_in_bounds[i]) & (x_in_bounds[i]):
-            ul = data[y_floor[i], x_floor[i]]
-            ur = data[y_floor[i], x_ceil[i]]
-            ll = data[y_ceil[i], x_floor[i]]
-            lr = data[y_ceil[i], x_ceil[i]]
-            value = ( ( ( 1 - tx[i] ) * ( 1 - ty[i] ) * ul )
-                    + ( tx[i] * ( 1 - ty[i] ) * ur )
-                    + ( ( 1 - tx[i] ) * ty[i] * ll )
-                    + ( tx[i] * ty[i] * lr ) )
-            out.flat[i] = value
-    return out
-
-@njit(UniTuple(float64[:], 2)(UniTuple(float64, 9), float64[:], float64[:]), parallel=True)
-def affine_map(affine, x, y):
-    a, b, c, d, e, f, _, _, _ = affine
-    n = x.size
-    new_x = np.zeros(n, dtype=np.float64)
-    new_y = np.zeros(n, dtype=np.float64)
-    for i in prange(n):
-        new_x[i] = x[i] * a + y[i] * b + c
-        new_y[i] = x[i] * d + y[i] * e + f
-    return new_x, new_y
